+++ stream_alert/alert_processor/main.py
@@ -18,7 +18,7 @@ import json
 
 from stream_alert.alert_processor import LOGGER
 from stream_alert.alert_processor.helpers import validate_alert
-from stream_alert.alert_processor.outputs import get_output_dispatcher
+from stream_alert.alert_processor.outputs.output_base import StreamAlertOutput
 from stream_alert.shared import NORMALIZATION_KEY
 
 
@@ -101,7 +101,12 @@ def run(alert, region, function_name, config):
             continue
 
         # Retrieve the proper class to handle dispatching the alerts of this services
-        output_dispatcher = get_output_dispatcher(service, region, function_name, config)
+        output_dispatcher = StreamAlertOutput.get_output_dispatcher(
+            service,
+            region,
+            function_name,
+            config
+        )
 
         if not output_dispatcher:
             continue
