+++ stream_alert/alert_processor/outputs/aws.py
@@ -1,0 +1,307 @@
+"""
+Copyright 2017-present, Airbnb Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+   http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+from abc import abstractmethod
+from collections import OrderedDict
+from datetime import datetime
+import json
+import uuid
+
+import backoff
+from botocore.exceptions import ClientError
+import boto3
+
+from stream_alert.alert_processor import LOGGER
+from stream_alert.alert_processor.outputs.output_base import (
+    OutputDispatcher,
+    OutputProperty,
+    StreamAlertOutput
+)
+from stream_alert.shared.backoff_handlers import (
+    backoff_handler,
+    success_handler,
+    giveup_handler
+)
+
+
+class AWSOutput(OutputDispatcher):
+    """Subclass to be inherited from for all AWS service outputs"""
+    def format_output_config(self, service_config, values):
+        """Format the output configuration for this AWS service to be written to disk
+
+        AWS services are stored as a dictionary within the config instead of a list so
+        we have access to the AWS value (arn/bucket name/etc) for Terraform
+
+        Args:
+            service_config (dict): The actual outputs config that has been read in
+            values (OrderedDict): Contains all the OutputProperty items for this service
+
+        Returns:
+            dict{<string>: <string>}: Updated dictionary of descriptors and
+                values for this AWS service needed for the output configuration
+            NOTE: S3 requires the bucket name, not an arn, for this value.
+                Instead of implementing this differently in subclasses, all AWSOutput
+                subclasses should use a generic 'aws_value' to store the value for the
+                descriptor used in configuration
+        """
+        return dict(service_config.get(self.__service__, {}),
+                    **{values['descriptor'].value: values['aws_value'].value})
+
+    @abstractmethod
+    def dispatch(self, **kwargs):
+        """Placeholder for implementation in the subclasses"""
+        pass
+
+
+@StreamAlertOutput
+class KinesisFirehoseOutput(AWSOutput):
+    """High throughput Alert delivery to AWS S3"""
+    MAX_RECORD_SIZE = 1000 * 1000
+    MAX_BACKOFF_ATTEMPTS = 3
+
+    __service__ = 'aws-firehose'
+    __aws_client__ = None
+
+    def get_user_defined_properties(self):
+        """Properties asssigned by the user when configuring a new Firehose output
+
+        Every output should return a dict that contains a 'descriptor' with a description of the
+        integration being configured.
+
+        Returns:
+            OrderedDict: Contains various OutputProperty items
+        """
+        return OrderedDict([
+            ('descriptor',
+             OutputProperty(
+                 description='a short and unique descriptor for this Firehose Delivery Stream')),
+            ('aws_value',
+             OutputProperty(description='the Firehose Delivery Stream name'))
+        ])
+
+    def dispatch(self, **kwargs):
+        """Send alert to a Kinesis Firehose Delivery Stream
+
+        Keyword Args:
+            descriptor (str): Service descriptor (ie: slack channel, pd integration)
+            rule_name (str): Name of the triggered rule
+            alert (dict): Alert relevant to the triggered rule
+
+        Returns:
+            bool: Indicates a successful or failed dispatch of the alert
+        """
+        @backoff.on_exception(backoff.fibo,
+                              ClientError,
+                              max_tries=self.MAX_BACKOFF_ATTEMPTS,
+                              jitter=backoff.full_jitter,
+                              on_backoff=backoff_handler,
+                              on_success=success_handler,
+                              on_giveup=giveup_handler)
+        def _firehose_request_wrapper(json_alert, delivery_stream):
+            """Make the PutRecord request to Kinesis Firehose with backoff
+
+            Args:
+                json_alert (str): The JSON dumped alert body
+                delivery_stream (str): The Firehose Delivery Stream to send to
+
+            Returns:
+                dict: Firehose response in the format below
+                    {'RecordId': 'string'}
+            """
+            return self.__aws_client__.put_record(DeliveryStreamName=delivery_stream,
+                                                  Record={'Data': json_alert})
+
+        if self.__aws_client__ is None:
+            self.__aws_client__ = boto3.client('firehose', region_name=self.region)
+
+        json_alert = json.dumps(kwargs['alert'], separators=(',', ':')) + '\n'
+        if len(json_alert) > self.MAX_RECORD_SIZE:
+            LOGGER.error('Alert too large to send to Firehose: \n%s...', json_alert[0:1000])
+            return False
+
+        delivery_stream = self.config[self.__service__][kwargs['descriptor']]
+        LOGGER.info('Sending alert [%s] to aws-firehose:%s',
+                    kwargs['rule_name'],
+                    delivery_stream)
+
+        resp = _firehose_request_wrapper(json_alert, delivery_stream)
+
+        if resp.get('RecordId'):
+            LOGGER.info('Alert [%s] successfully sent to aws-firehose:%s with RecordId:%s',
+                        kwargs['rule_name'],
+                        delivery_stream,
+                        resp['RecordId'])
+
+        return self._log_status(resp)
+
+
+@StreamAlertOutput
+class S3Output(AWSOutput):
+    """S3Output handles all alert dispatching for AWS S3"""
+    __service__ = 'aws-s3'
+
+    def get_user_defined_properties(self):
+        """Get properties that must be asssigned by the user when configuring a new S3
+        output.  This should be sensitive or unique information for this use-case that needs
+        to come from the user.
+
+        Every output should return a dict that contains a 'descriptor' with a description of the
+        integration being configured.
+
+        S3 also requires a user provided bucket name to be used for this service output. This
+        value should not be masked during input and is not a credential requirement
+        that needs encrypted.
+
+        Returns:
+            OrderedDict: Contains various OutputProperty items
+        """
+        return OrderedDict([
+            ('descriptor',
+             OutputProperty(
+                 description='a short and unique descriptor for this S3 bucket (ie: bucket name)')),
+            ('aws_value',
+             OutputProperty(description='the AWS S3 bucket name to use for this S3 configuration'))
+        ])
+
+    def dispatch(self, **kwargs):
+        """Send alert to an S3 bucket
+
+        Organizes alert into the following folder structure:
+            service/entity/rule_name/datetime.json
+        The alert gets dumped to a JSON string
+
+        Args:
+            **kwargs: consists of any combination of the following items:
+                descriptor (str): Service descriptor (ie: slack channel, pd integration)
+                rule_name (str): Name of the triggered rule
+                alert (dict): Alert relevant to the triggered rule
+        """
+        alert = kwargs['alert']
+        service = alert['source_service']
+        entity = alert['source_entity']
+
+        current_date = datetime.now()
+
+        s3_alert = alert
+        # JSON dump the alert to retain a consistent alerts schema across log types.
+        # This will get replaced by a UUID which references a record in a
+        # different table in the future.
+        s3_alert['record'] = json.dumps(s3_alert['record'])
+        alert_string = json.dumps(s3_alert)
+
+        bucket = self.config[self.__service__][kwargs['descriptor']]
+
+        # Prefix with alerts to account for generic non-streamalert buckets
+        # Produces the following key format:
+        #   alerts/dt=2017-01-25-00/kinesis_my-stream_my-rule_uuid.json
+        # Keys need to be unique to avoid object overwriting
+        key = 'alerts/dt={}/{}_{}_{}_{}.json'.format(
+            current_date.strftime('%Y-%m-%d-%H'),
+            service,
+            entity,
+            alert['rule_name'],
+            uuid.uuid4()
+        )
+
+        LOGGER.debug('Sending alert to S3 bucket %s with key %s', bucket, key)
+
+        client = boto3.client('s3', region_name=self.region)
+        resp = client.put_object(Body=alert_string,
+                                 Bucket=bucket,
+                                 Key=key)
+
+        return self._log_status(resp)
+
+
+@StreamAlertOutput
+class LambdaOutput(AWSOutput):
+    """LambdaOutput handles all alert dispatching to AWS Lambda"""
+    __service__ = 'aws-lambda'
+
+    def get_user_defined_properties(self):
+        """Get properties that must be asssigned by the user when configuring a new Lambda
+        output.  This should be sensitive or unique information for this use-case that needs
+        to come from the user.
+
+        Every output should return a dict that contains a 'descriptor' with a description of the
+        integration being configured.
+
+        Sending to Lambda also requires a user provided Lambda function name and optional qualifier
+        (if applicabale for the user's use case). A fully-qualified AWS ARN is also acceptable for
+        this value. This value should not be masked during input and is not a credential requirement
+        that needs encrypted.
+
+        Returns:
+            OrderedDict: Contains various OutputProperty items
+        """
+        return OrderedDict([
+            ('descriptor',
+             OutputProperty(description='a short and unique descriptor for this Lambda function '
+                                        'configuration (ie: abbreviated name)')),
+            ('aws_value',
+             OutputProperty(description='the AWS Lambda function name, with the optional '
+                                        'qualifier (aka \'alias\'), to use for this '
+                                        'configuration (ie: output_function:qualifier)',
+                            input_restrictions={' '})),
+        ])
+
+    def dispatch(self, **kwargs):
+        """Send alert to a Lambda function
+
+        The alert gets dumped to a JSON string to be sent to the Lambda function
+
+        Args:
+            **kwargs: consists of any combination of the following items:
+                descriptor (str): Service descriptor (ie: slack channel, pd integration)
+                rule_name (str): Name of the triggered rule
+                alert (dict): Alert relevant to the triggered rule
+        """
+        alert = kwargs['alert']
+        alert_string = json.dumps(alert['record'])
+        function_name = self.config[self.__service__][kwargs['descriptor']]
+
+        # Check to see if there is an optional qualifier included here
+        # Acceptable values for the output configuration are the full ARN,
+        # a function name followed by a qualifier, or just a function name:
+        #   'arn:aws:lambda:aws-region:acct-id:function:function-name:prod'
+        #   'function-name:prod'
+        #   'function-name'
+        # Checking the length of the list for 2 or 8 should account for all
+        # times a qualifier is provided.
+        parts = function_name.split(':')
+        if len(parts) == 2 or len(parts) == 8:
+            function = parts[-2]
+            qualifier = parts[-1]
+        else:
+            function = parts[-1]
+            qualifier = None
+
+        LOGGER.debug('Sending alert to Lambda function %s', function_name)
+
+        client = boto3.client('lambda', region_name=self.region)
+        # Use the qualifier if it's available. Passing an empty qualifier in
+        # with `Qualifier=''` or `Qualifier=None` does not work and thus we
+        # have to perform different calls to client.invoke().
+        if qualifier:
+            resp = client.invoke(FunctionName=function,
+                                 InvocationType='Event',
+                                 Payload=alert_string,
+                                 Qualifier=qualifier)
+        else:
+            resp = client.invoke(FunctionName=function,
+                                 InvocationType='Event',
+                                 Payload=alert_string)
+
+        return self._log_status(resp)
